backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
ply==3.11
lark==1.2.2
python-multipart==0.0.20
python-dotenv==1.0.1